requirements.txt
Django>=5.2.4
djangorestframework
djangorestframework-simplejwt
drf-yasg